change xChart version to 2.6.0 to 2.6.1
xChart depends on VectorGraphics2D. xChart 2.6.0 depended on VectorGraphics2D 0.9.2 in its respective pom.xml file. But 0.9.2 does not exist in maven central anymore. The oldest version of VectorGraphics2D on maven central is 0.9.3. To solve this problem, upgrade xChart to 2.6.1. xChart 2.6.1 depends on VectorGraphics2D 0.9.3, which exists where maven looks for it. So, maven can build riot-bench successfully.